+++ examples/components/Form.js
@@ -1,7 +1,29 @@
 import React, { useCallback, useState } from 'react';
 
-import JoditEditor from '../../src/';
+import JoditEditor, { Jodit } from '../../src/';
 import './Form.css';
+
+/**
+ * @param {Jodit} jodit
+ */
+function preparePaste(jodit) {
+	jodit.e.on(
+		'paste',
+		e => {
+			if (confirm('Change pasted content?')) {
+				jodit.e.stopPropagation('paste');
+				jodit.s.insertHTML(
+					Jodit.modules.Helpers.getDataTransfer(e)
+						.getData(Jodit.constants.TEXT_HTML)
+						.replace(/a/g, 'b')
+				);
+				return false;
+			}
+		},
+		{ top: true }
+	);
+}
+Jodit.plugins.add('preparePaste', preparePaste);
 
 const Form = () => {
 	const [isSource, setSource] = useState(false);