+++ src/JoditEditor.d.ts
@@ -1,13 +1,16 @@
 import * as React from 'react';
-import { IJodit } from 'jodit';
+import { Jodit } from 'jodit/types/jodit';
 
 declare module 'jodit-react' {
 	export interface IJoditEditorProps {
-		value: string,
-		config?: Partial<IJodit['options']>;
+		value: string;
+		config?: Partial<Jodit['options']>;
 		onChange: (newValue: string) => void;
 		onBlur: (newValue: string) => void;
 	}
-	const JoditEditor: React.ComponentType<IJoditEditorProps>;
+	const JoditEditor: React.ForwardRefExoticComponent<
+		React.PropsWithoutRef<IJoditEditorProps> & React.RefAttributes<Jodit>
+	>;
 	export default JoditEditor;
+	export { Jodit };
 }