+++ src/JoditEditor.js
@@ -95,5 +95,3 @@ JoditEditor.propTypes = {
 };
 
 export default JoditEditor;
-
-export { Jodit };