+++ src/index.js
@@ -1,3 +1,5 @@
 import JoditEditor from './JoditEditor';
+import { Jodit } from './include.jodit';
 
 export default JoditEditor;
+export { Jodit };